fix: Error flash before contract not found shown.
